requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
urllib3==2.0.4
yamale==4.0.4
